+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/launcher/LauncherController.java
@@ -37,16 +37,11 @@ import org.slf4j.LoggerFactory;
 
 import java.awt.*;
 import java.io.File;
-import java.net.URI;
-import java.net.URL;
-import java.util.ArrayList;
+import java.net.*;
+import java.nio.charset.StandardCharsets;
+import java.util.*;
 import java.util.List;
-import java.util.Optional;
-import java.util.ResourceBundle;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
-import java.util.concurrent.TimeUnit;
+import java.util.concurrent.*;
 
 import static de.mephisto.vpin.ui.Studio.client;
 
@@ -98,6 +93,8 @@ public class LauncherController implements Initializable {
   private Stage stage;
   private PropertiesStore store;
 
+  private final Map<InetAddress, BroadcastInfo> broadcastData = new ConcurrentHashMap<>();
+
   @FXML
   private void onHelp() {
     if (Desktop.getDesktop().isSupported(Desktop.Action.BROWSE)) {
@@ -105,7 +102,7 @@ public class LauncherController implements Initializable {
         Desktop.getDesktop().browse(new URI("https://github.com/syd711/vpin-studio/wiki/Troubleshooting"));
       }
       catch (Exception ex) {
-        LOG.error("Failed to open link: " + ex.getMessage(), ex);
+        LOG.error("Failed to open link: {}", ex.getMessage(), ex);
       }
     }
   }
@@ -117,7 +114,7 @@ public class LauncherController implements Initializable {
       dotNetInstalled = WinRegistry.isDotNetInstalled();
     }
     catch (Exception e) {
-      LOG.error("Error checking .net framework: " + e.getMessage(), e);
+      LOG.error("Error checking .net framework: {}", e.getMessage(), e);
       WidgetFactory.showAlert(stage, "Error", "Error checking .net framework: " + e.getMessage());
     }
 
@@ -149,9 +146,23 @@ public class LauncherController implements Initializable {
 
       List<Object> entries = store.getEntries();
       for (Object entry : entries) {
-        LOG.info("Checking connection to " + entry);
+        LOG.info("Checking connection to {}", entry);
         connection = checkConnection(String.valueOf(entry));
         if (connection != null) {
+          result.add(connection);
+        }
+      }
+
+      for (Map.Entry<InetAddress, BroadcastInfo> entry : broadcastData.entrySet()) {
+        InetAddress ip = entry.getKey();
+        BroadcastInfo info = entry.getValue();
+        String systemName = info.getSystemName();
+
+        LOG.info("Checking broadcasted connection to {} with system name {}", ip.getHostAddress(), systemName);
+        connection = checkConnection(ip.getHostAddress());
+        VPinConnection finalConnection = connection;
+        if (connection != null && result.stream().noneMatch(conn -> conn.getHost().equals(finalConnection.getHost()))) {
+          connection.setDiscovered(true);
           result.add(connection);
         }
       }
@@ -233,6 +244,13 @@ public class LauncherController implements Initializable {
   public void setStage(Stage stage) {
     this.stage = stage;
     this.onConnectionRefresh();
+
+    // Start listening when the window opens
+    this.startBroadcastListener();
+
+    // Stop the listener when the window is closed
+    stage.setOnHidden(event -> this.stopBroadcastListener());
+    stage.setOnCloseRequest(event -> this.stopBroadcastListener());
   }
 
   private void installServer() {
@@ -259,7 +277,7 @@ public class LauncherController implements Initializable {
           }
         }
 
-        LOG.info("Found server startup, running on version " + client.getSystemService().getVersion() + ", starting table scan.");
+        LOG.info("Found server startup, running on version {}, starting table scan.", client.getSystemService().getVersion());
         Platform.runLater(() -> {
           stage.close();
           ProgressDialog.createProgressDialog(new ServiceInstallationProgressModel(Studio.client));
@@ -268,7 +286,7 @@ public class LauncherController implements Initializable {
       }).start();
     }
     catch (Exception e) {
-      LOG.error("Server installation failed: " + e.getMessage(), e);
+        LOG.error("Server installation failed: {}", e.getMessage(), e);
       WidgetFactory.showAlert(stage, "Server installation failed.", "Error: " + e.getMessage());
     }
   }
@@ -313,26 +331,46 @@ public class LauncherController implements Initializable {
       return new SimpleObjectProperty(view);
     });
 
-    actionColumn.setCellValueFactory(cellData -> {
-      VPinConnection value = cellData.getValue();
-      if (value.getHost().equals("localhost")) {
-        return new SimpleObjectProperty<>("");
-      }
+    actionColumn.setCellFactory(col -> new TableCell<>() {
+      @Override
+      protected void updateItem(String item, boolean empty) {
+        super.updateItem(item, empty);
+        setText(null);
 
-      Button button = new Button();
-      button.setStyle("-fx-border-radius: 6px;");
-      FontIcon icon = new FontIcon("mdi2d-delete-outline");
-      icon.setIconSize(8);
-      icon.setIconColor(Paint.valueOf("#FFFFFF"));
-      button.setGraphic(icon);
-      button.setOnAction(event -> {
-        Optional<ButtonType> result = WidgetFactory.showConfirmation(stage, "Delete connection to '" + value + "'?");
-        if (result.isPresent() && result.get().equals(ButtonType.OK)) {
-          store.removeValue(String.valueOf(value.getHost()));
-          onConnectionRefresh();
+        if (empty || getTableRow() == null || getTableRow().getItem() == null) {
+          setGraphic(null);
+          return;
         }
-      });
-      return new SimpleObjectProperty(button);
+
+        VPinConnection value = getTableRow().getItem();
+
+        // No icon for localhost
+        if (value.getHost().equals("localhost")) {
+          setGraphic(null);
+        } else if (value.getDiscovered()) {
+          // Show a custom icon for discovered connections
+          FontIcon discoveredIcon = new FontIcon("mdi2a-antenna");
+          discoveredIcon.setIconSize(16);
+          discoveredIcon.setIconColor(Paint.valueOf("#2196F3"));
+          setGraphic(discoveredIcon);
+        } else {
+          // Show the delete button for non-discovered connections
+          Button button = new Button();
+          button.setStyle("-fx-border-radius: 6px;");
+          FontIcon deleteIcon = new FontIcon("mdi2d-delete-outline");
+          deleteIcon.setIconSize(8);
+          deleteIcon.setIconColor(Paint.valueOf("#FFFFFF"));
+          button.setGraphic(deleteIcon);
+          button.setOnAction(event -> {
+            Optional<ButtonType> result = WidgetFactory.showConfirmation(stage, "Delete connection to '" + value.getHost() + "'?");
+            if (result.isPresent() && result.get().equals(ButtonType.OK)) {
+              store.removeValue(String.valueOf(value.getHost()));
+              onConnectionRefresh();
+            }
+          });
+          setGraphic(button);
+        }
+      }
     });
 
     tableView.setRowFactory(tv -> {
@@ -389,5 +427,145 @@ public class LauncherController implements Initializable {
     }
 
     return null;
+  }
+
+  private Thread broadcastListenerThread;
+  private Thread staleEntriesRemoverThread;
+  private DatagramSocket socket;
+  private boolean shouldListen = false;
+
+  private void listenForBroadcast() {
+    try {
+      socket = new DatagramSocket(50505);
+      byte[] buffer = new byte[256];
+      DatagramPacket packet = new DatagramPacket(buffer, buffer.length);
+
+      while (shouldListen) {
+        socket.receive(packet);
+
+        String receivedSystemName = new String(packet.getData(), 0, packet.getLength(), StandardCharsets.US_ASCII);
+        if (receivedSystemName.isEmpty()) {
+          receivedSystemName = UIDefaults.VPIN_NAME;
+        }
+
+        InetAddress senderAddress = packet.getAddress();
+
+        long currentTime = System.currentTimeMillis();
+        if (!broadcastData.containsKey(senderAddress)) {
+          // Log the received information
+            LOG.info("Received broadcast for the first time from IP: {}, System Name: {}", senderAddress.getHostAddress(), receivedSystemName);
+
+          // Store the IP and system name
+          broadcastData.put(senderAddress, new BroadcastInfo(receivedSystemName, currentTime));
+
+          // Notify that broadcast data has changed
+          onBroadcastDataChanged();
+        } else {
+          BroadcastInfo existingInfo = broadcastData.get(senderAddress);
+          existingInfo.updateLastBroadcastTime(currentTime);
+        }
+      }
+    } catch (Exception e) {
+        LOG.error("Error receiving broadcast: {}", e.getMessage(), e);
+    }
+  }
+
+  public synchronized void startBroadcastListener() {
+    if (broadcastListenerThread != null) {
+      LOG.info("Broadcast listener already started");
+      return;
+    }
+
+    shouldListen = true;
+
+    broadcastListenerThread = new Thread(this::listenForBroadcast);
+
+    broadcastListenerThread.start();
+
+    // Start the thread that removes stale entries
+    startStaleEntriesRemover();
+  }
+
+  public synchronized void stopBroadcastListener() {
+    if (broadcastListenerThread == null) {
+      LOG.info("Broadcast listener not running");
+      return;
+    }
+
+    shouldListen = false;
+
+    try {
+      broadcastListenerThread.join(); // Wait for the listener thread to stop
+    } catch (InterruptedException e) {
+      LOG.error("Failed to stop broadcast listener: {}", e.getMessage(), e);
+    }
+
+    // Close the DatagramSocket to release the port
+    if (socket != null && !socket.isClosed()) {
+      socket.close();
+    }
+
+    LOG.info("Broadcast listener stopped");
+    broadcastListenerThread = null;
+
+    stopStaleEntriesRemover();
+  }
+
+  private void onBroadcastDataChanged() {
+    Platform.runLater(this::onConnectionRefresh);
+  }
+
+  public synchronized void startStaleEntriesRemover() {
+    if (staleEntriesRemoverThread != null) {
+      LOG.info("Stale entries remover already started");
+      return;
+    }
+
+    staleEntriesRemoverThread = new Thread(() -> {
+      while (shouldListen) { // Stop checking if shouldListen is false
+        long currentTime = System.currentTimeMillis();
+        boolean dataChanged = false;
+
+        // Iterate over the broadcastData and remove stale entries
+        for (InetAddress ip : new ArrayList<>(broadcastData.keySet())) {
+          BroadcastInfo info = broadcastData.get(ip);
+          if (currentTime - info.getLastBroadcastTime() > 10000) { // More than 10 seconds have passed
+              LOG.info("Removing stale broadcast data for IP: {}", ip.getHostAddress());
+            broadcastData.remove(ip);
+            dataChanged = true;
+          }
+        }
+
+        if (dataChanged) {
+          onBroadcastDataChanged();
+        }
+
+        try {
+          Thread.sleep(2000); // Check every 2 seconds
+        } catch (InterruptedException e) {
+          Thread.currentThread().interrupt();
+        }
+      }
+    });
+
+    staleEntriesRemoverThread.start();
+  }
+
+  public synchronized void stopStaleEntriesRemover() {
+    if (staleEntriesRemoverThread == null) {
+      LOG.info("Stale entries remover not running");
+      return;
+    }
+
+    shouldListen = false; // This will stop the stale entries remover loop
+
+    try {
+      staleEntriesRemoverThread.join(); // Wait for the stale entries remover thread to stop
+    } catch (InterruptedException e) {
+        LOG.error("Failed to stop stale entries remover: {}", e.getMessage(), e);
+    }
+
+    LOG.info("Stale entries remover stopped");
+    staleEntriesRemoverThread = null;
   }
 }